:truck: move shared model to specific contexte
