+++ api/tests/certification/evaluation/integration/infrastructure/repositories/certifiable-profile-for-learning-content-repository_test.js
@@ -1,5 +1,5 @@
+import { CertifiableProfileForLearningContent } from '../../../../../../src/certification/evaluation/domain/models/CertifiableProfileForLearningContent.js';
 import * as certifiableProfileForLearningContentRepository from '../../../../../../src/certification/evaluation/infrastructure/repositories/certifiable-profile-for-learning-content-repository.js';
-import { CertifiableProfileForLearningContent } from '../../../../../../src/shared/domain/models/CertifiableProfileForLearningContent.js';
 import { KnowledgeElement } from '../../../../../../src/shared/domain/models/KnowledgeElement.js';
 import { databaseBuilder, domainBuilder, expect } from '../../../../../test-helper.js';
